+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial strategy: Partitioning Around Medoids
 - Initial strategy: K-means
 - Initial strategy: Gaussian Mixture Model
+- Constraints and conditions for discrete parameters.
 
 ### Fixed
 - UCB now usable as acquisition function, hard-set beta parameter to 1.0.